+++ NEWS.md
@@ -8,6 +8,7 @@
 - **BREAKING CHANGE** The `includeShinyjs` parameter in `runcodeUI()` is now deprecated
 - Fix bug: shinyjs functions did not work with IDs that had a space in them; fixes #176
 - Fix bug #186: non-selectize select inputs could not be disabled
+- Added `id` parameter to `runcode()`, allowing it to work inside Shiny modules (#184)
 
 # shinyjs 1.0